drivers: uart_aspeed: Add interrupt driven support
Refactor ASPEED UART driver to support the interrupt
driven operation.

It is NOT recommended to apply UDMA on the UART device
used for the interactive shell.

This is due to the interrupt behavior of UDMA differs
from that of generic ns16550 device. As Zephyr shell
implementation heavily relies on the UART interrupt
to handle TX/RX bytes, the difference will result in
the console slow down to a crawl.

Signed-off-by: Chia-Wei Wang <chiawei_wang@aspeedtech.com>
Change-Id: Ibedca63e4baf5683d0de9ab62c4f575ec0b49d78