soc: ast10x0: Add gpio passthrough pinctrl setting.
The gpio passthrough is used to redirect the input of one gpio to the
output of another gpio.
In ast10x0, we support 4 sets of passthrough pin pairs, listed below:
GPIOC0 → GPIOC1
GPIOC2 → GPIOC3
GPIOC4 → GPIOC5
GPIOC6 → GPIOC7
This patch can add pinctrl-0 to the dts to enable this feature.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I740b6d086071cf845cead8f4f322c0cc76f10267