i3c: aspeed: Add the option to enable the pec for priv-xfer.
Enable the priv-xfer-pec property when i3c want to communicate with data
that have PEC. The PEC will auto append to the tail of the data when doing
private transfer and verify the PEC when receiving the data from master.
The reason for using software pec instead of the hardware pec is that
latter can't distinguish whether the PEC is used in CCC or private
transfer and the controller can't aware DEVCTRL CCC to decide the time to
enable/disable the PEC function.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I8762c7d323fc04d29afd4076597c2baf117501aa