soc: arm: ast26xx: fix system crash when SSP reset
The SSP will crash after reset assert and deassert without reloading the
image.  This is due to the dirty .data section.  Enable CONFIG_XIP to
let the VMA of .data section be different with the LMA.  Which means:

.data section LMA --> Read Only region of the SRAM
.data section VMA --> Read/Write region of the SRAM

The .data section will be copied from LMA to VMA during prep_c.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: I6e54982409c8d9233c841cafc1e77783317b55f1