drivers: gpio: aspeed: Fix the pin value for single ended mode.
When setting the flag as GPIOF_OPEN_DRAIN then it will assume that pins is
open drain type. Such pins will not be driven to 1 in output mode. It is
require to connect pull-up on such pins. The pin is make to LOW by driving
value 0 in output mode.

When setting the flag as GPIOF_OPEN_SOURCE then it will assume that pins is
open source type. Such pins will not be driven to 0 in output mode. It is
require to connect pull-down on such pin. The pin is make to HIGH by
driving value 1 in output mode.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I7e03b2c6cafae67c0e56938867268b809eabf4ea