i3c: aspeed: Add software flag for allowing SIR.
The hardware flag sir_allowed has a default value of 1, and it cannot be
cleared by the software, causing the software to lack a suitable flag to
determine when the master is ready to receive the IBI.
In order to solve this issue, a new boolean variable sir_allowed_by_sw
should be introduced in the i3c_aspeed_init() function, with an initial
value of 0. Furthermore, a new worker function called sir_allowed_worker()
should also be incorporated into the system. This worker will be submitted
only after the dynamic address of the slave has been assigned. Its primary
responsibility is to set the value of the sir_allowed_by_sw variable 1
after a delay of 1second which is sufficient for the i3c master to
complete the bus initialization process.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I0668e705bf394882a6e587d8acd7c710e5213989