i3c: aspeed: Disable static addressing if it is not required.
Skip setting the static address if the "assigned-address" property is not
assigned in slave mode. This patch is intended to prevent a valid static
address from affecting the behavior of the i3c slave. The static address
should only be enabled when the master wants to assign a dynamic address
using setaasa or setdasa, or when the slave wants to operate in I2C mode.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I285b841d163b866f5865707b5d8609d6011ac415