i3c: aspeed: Remove the public api for PID and static address.
The static address and PID should not be changed after enabling the i3c
slave controller to avoid any synchronization issues with the master
during address assignment flow. It is important to ensure that any
necessary configuration of the static address and PID is completed
before enabling the i3c slave controller.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I6f5811c3730bb6b723fbed0148544944b7e7cebb