adc: aspeed: Support deglitch feature.
Add the parameters upper_bound, lower_bound, and deglitch_en for ADC
channel configuration. When deglitch_en is set to true, the driver will
use the upper_bound and lower_bound as threshold values. If the ADC value
falls outside this threshold, the driver will wait for the ADC sampling
period and perform an additional read once to achieve the deglitching
purpose.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: Id96e29f6851bd937caed3ee09c49cb31e88dcaf8