adc: aspeed: Always apply the cv for adc read.
This patch ensures that the compensating value is consistently applied to
ADC readings, thereby enhancing the accuracy of the ADC measurements.
Furthermore, it is necessary to fix the logic regarding the "calibrate"
parameter in the ADC reading sequence. Instead of using it to determine
whether to apply the CV, it should be utilized to determine whether to
rerun the calibration process for getting new cv.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: Id86a150ec3765e70b3db5fa5cdcce22b743c977f