drivers:i3c: fix calling unregistered callback function
It set the data type of pos "uint32_t".
However, i3c_aspeed_get_pos function return status is "int".
If the slave address did not be found in i3c_aspeed_get_pos
function, this function returned "-1". Due to the incorrect
data type of pos, the value of pos would be "0xffffffff"

The max i3c device descriptor was 32, so it got the NULL
i3c device descriptor and called unregistered callback
function and it caused system crash.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: I9a49a9f784526c13c73b103ff8261cb8678a7c75