i3c: aspeed: Reorganize the APIs for entering and exiting the HALT state
In certain situations, such as receiving an error response or
experiencing a timeout during transfer transmission, it becomes
necessary to reset the hardware queues. This reset operation must be
executed when the controller is in the HALT state.

To achieve this, this commit adds the `enter_halt` API to handle these
two cases:
- Receiving an error response: the I3C controller will automatically
enter the HALT state. The API will monitore `present_state` register for
this.

- Timeout during transfer transmission: Use the `abort` bit to force the
controller to enter the HALT state.

When the reset is done, use `exit_halt` API to exit from the HALT state
and back to the normal operation state.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: Icb63e5cfd5732c2e2b16a1752455673f577cac46