soc: aspeed: Declare the non-cached memory in DTS
The non-cached memory region was creating by manipulating the linker
script, making it difficult to read and comprehend. This commit
introduces `LINKER_DT*` macros to facilitate the declaration of the
non-cached memory region in the DTS using the property
`zephyr,memory-region="RAM_NC"`. To allocate sections in the RAM_NC,
simply append the attribute `__section("RAM_NC")` to the varibles.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: Ic7024e3600eddb18a2eaaa838f598b6c04454844