i3c: aspeed: Add api to hook the reset callback
Currently, when the I3C slave encounters a transfer error, it
automatically enters a halt state. After that, the hardware lacks a
recovery flow to restore the controller, so the software has to reset the
I3C controller. To allow users to be aware of this reset event, this patch
provides an API to hook into the reset callback. This callback can be used
to execute a hot-join request or any user-defined behavior to reattach the
I3C slave to the bus, ensuring that future transfers can continue without
issues.

Example:
void i3c_rst_cb_test(const struct device *dev)
{
	i3c_slave_hj_req(dev);
}

void main()
{
	const struct device *dev;
	dev = device_get_binding("I3C_0");
	...
	i3c_hook_rst_cb(dev, i3c_rst_cb_test);
}

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: Ia9082183ebf0ceb38baae7d437f64588500cf0a1