i3c: aspeed: Avoid to reinitialize the work
When the I3C controller encounters an error that requires a reset and the
aspeed_i3c_init function is called to reinitialize the controller,
reinitializing the work at this moment may cause a racing issue. This
happens because the work could be in use simultaneously, leading to a
kernel panic.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I2074019f10f403c504757342a353b352d28b3bbc