drivers:cache_ast27xx: Force all memory area be cached
The SSP cache controller includes a pre-fetch buffer that is not
directly configurable by software. When the cache area control register
are set to partially cacheable and partially non-cacheable, the
pre-fetch buffer does not function as expected, resulting in increased
latency for cacheable memory access. To restore normal latency,
configure SCU0_12C and SCU0_130 to 0xffffffff, ensuring that all
instruction and data memory accesses go through the cache.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: I7386ca69fab96aad796a73c2ff244f711814a86a