drivers: ipm: ast2700: Fix wait with IPCR_STATUS check
For the wait usage, if nonzero, busy-wait for remote to
consume the message. The message is considered consumed
once the remote interrupt handler finishes. If there is
deferred processing on the remote side, or you would like
to queue outgoing messages and wait on an event/semaphore,
you can implement that in a high-level driver

Change-Id: I5bd67611af208de7ee8dab07856f0901a92ac9a2
Signed-off-by: Kevin Chen <kevin_chen@aspeedtech.com>