drivers/ltpi: Ensure 25M speed capability is preserved
The LTPI specification mandates that the lowest speed (25M) must always
be supported. Previously, the code only checked if `phy_speed_cap` was
zero before setting the 25M capability. This fails to preserve other
speed capabilities if they are already set.

Update the condition to only set the 25M bit if it is not already
present, allowing other capabilities to coexist.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: I7fd40d1bee2a09f5ebd85aa86f724356109cfa87