arch/arm: Add support for relocating vector table to SRAM
Enables relocating the vector table to SRAM when configured with
CONFIG_SRAM_VECTOR_TABLE. Adds logic to copy the vector table from
flash to SRAM and updates the vector table address in SCB->VTOR.

Updates the linker scripts to place the vector table at the beginning
of SRAM when the relevant configuration is enabled.

Facilitates scenarios requiring vector tables in SRAM for specific
runtime conditions.

Change-Id: Ia85d068a66e8ef8289301a2fb5bc3a8b7165fa6e
Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>