boards: add support for AST2700 SSP-TSP variant
Add a new SSP-TSP variant to the AST2700 EVB board configuration. This
variant is used to indicate that the SSP is built with a TSP present, so
the SRAM size should be shared.
The SRAM sharing logic follows the formula below:
SSP: 24 KB (reserved for IPC)
TSP: 8 KB (reserved for stack, The stack of the SSP in TCM)
The remaining memory is equally divided between SSP and TSP:
(128 - 24 - 8) / 2 = 48KB

Change-Id: I0149d9bf6ae5da5a298061880857c51c61d876cf
Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>